Move UInt64Tests
