+++ Tests/Batch4/Batch4.csproj
@@ -67,7 +67,6 @@
     <Compile Include="Text\RegularExpressions\RegexTests.cs" />
     <Compile Include="SimpleTypes\UInt16Tests.cs" />
     <Compile Include="SimpleTypes\UInt32Tests.cs" />
-    <Compile Include="SimpleTypes\UInt64Tests.cs" />
     <Compile Include="SimpleTypes\SByteTests.cs" />
     <Compile Include="SimpleTypes\Int32Tests.cs" />
     <Compile Include="SimpleTypes\Int16Tests.cs" />
